requirements.txt
mcp[cli]
fastmcp
chromadb
sentence-transformers
pyyaml
